+++ tests/testthat/test-paired_test_categorical.R
@@ -20,3 +20,40 @@ test_that("output structure is correct", {
       names(paired_test_categorical(tab_more))
   ))
 })
+
+test_that("wrong input return NA list", {
+  expect_is(paired_test_categorical(c(1, 2)), "list")
+  expect(is.na(paired_test_categorical(c(1, 2))[["P"]]))
+
+  expect_is(paired_test_categorical(matrix(c(1, 2))), "list")
+  expect(is.na(paired_test_categorical(matrix(c(1, 2)))[["P"]]))
+
+  expect_is(
+    paired_test_categorical(matrix(c(1, 2), ncol = 2)),
+    "list"
+  )
+  expect(
+    is.na(paired_test_categorical(matrix(c(1, 2), ncol = 2))[["P"]])
+  )
+
+})
+
+
+test_that("singular matrix were managed", {
+  expect_warning(
+    paired_test_categorical(matrix(c(1, 0, 2, 0), ncol = 2))
+  )
+
+  expect_is(
+    suppressWarnings(
+      paired_test_categorical(matrix(c(1, 0, 2, 0), ncol = 2))
+    ),
+    "list"
+  )
+  expect(is.na(
+    suppressWarnings(
+      paired_test_categorical(matrix(c(1, 0, 2, 0), ncol = 2))[["P"]]
+    )
+  ))
+
+})